backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
aiofiles==23.2.1
langgraph==0.1.19
langchain==0.2.6
langchain-google-genai==1.0.6
moviepy==1.0.3
opencv-python==4.10.0.84
ultralytics==8.2.50
pydub==0.25.1
TTS==0.22.0
pdfplumber==0.11.1
python-docx==1.1.2
numpy==1.26.4
pillow==10.3.0
httpx==0.27.0
python-dotenv==1.0.1
google-generativeai==0.5.4
python-jose==3.3.0
